+++ statham/dsl/elements/meta.py
@@ -85,25 +85,43 @@ class ObjectMeta(type, Element):
         minProperties: Maybe[int] = NotPassed(),
         maxProperties: Maybe[int] = NotPassed(),
         patternProperties: Maybe[Dict[str, Element]] = NotPassed(),
-        additionalProperties: Union[Element, bool] = True,
+        additionalProperties: Maybe[Union[Element, bool]] = NotPassed(),
         propertyNames: Maybe[Element] = NotPassed(),
         dependencies: Maybe[Dict[str, Union[List[str], Element]]] = NotPassed(),
     ):
         cls: ObjectMeta = cast(
             ObjectMeta, type.__new__(mcs, name, bases, dict(classdict))
         )
-        cls.default = default
-        cls.const = const
-        cls.enum = enum
-        cls.required = required
+        previous = lambda attr, default: getattr(cls, attr, default)
+        get_value = (
+            lambda value, attr: value
+            if not isinstance(value, NotPassed)
+            else previous(attr, NotPassed())
+        )
+        cls.default = get_value(default, "default")
+        cls.const = get_value(const, "const")
+        cls.enum = get_value(enum, "enum")
+        cls.required = get_value(required, "required")
         # https://github.com/python/mypy/issues/3004
-        cls.properties = classdict.properties  # type: ignore
-        cls.minProperties = minProperties
-        cls.maxProperties = maxProperties
-        cls.patternProperties = patternProperties
-        cls.additionalProperties = additionalProperties
-        cls.propertyNames = propertyNames
-        cls.dependencies = dependencies
+        cls.properties = {  # type: ignore
+            **{
+                attr: prop.clone()
+                for attr, prop in previous("properties", {}).items()
+            },
+            **classdict.properties,
+        }
+        cls.minProperties = get_value(minProperties, "minProperties")
+        cls.maxProperties = get_value(maxProperties, "maxProperties")
+        cls.patternProperties = get_value(
+            patternProperties, "patternProperties"
+        )
+        cls.additionalProperties = (
+            additionalProperties
+            if not isinstance(additionalProperties, NotPassed)
+            else previous("additionalProperties", True)
+        )
+        cls.propertyNames = get_value(propertyNames, "propertyNames")
+        cls.dependencies = get_value(dependencies, "dependencies")
         return cls
 
     def __hash__(cls):
@@ -145,7 +163,9 @@ class ObjectMeta(type, Element):
             if param.kind != param.KEYWORD_ONLY:
                 continue
             value = getattr(cls, param.name, NotPassed())
-            if value == param.default:
+            if value == param.default or (
+                param.name == "additionalProperties" and value is True
+            ):
                 continue
             cls_args.append(f"{param.name}={repr(value)}")
         class_def = f"""class {repr(cls)}({', '.join(cls_args)}):